per packet instead of quantity

+++ src/models/orderItem.model.ts
@@ -5,13 +5,13 @@ export class OrderItem extends Model {
   public id!: number;
   public orderId!: number;
   public productId!: number;
-  public quantity!: number;
-  public packages!: number; // Add package count
-  public originalPrice!: number;
-  public adjustedPrice!: number | null;
+  public quantity!: number; // Total pieces (calculated from packages × VPE)
+  public packages!: number; // Number of packages (can be fractional)
+  public originalPrice!: number; // Price per packet
+  public adjustedPrice!: number | null; // Adjusted price per packet
   public taxRate!: number; // Tax percentage
   public taxAmount!: number; // Calculated tax
-  public unitPerPackageSnapshot!: number;
+  public unitPerPackageSnapshot!: number; // VPE snapshot
   // Add association properties
   public order?: Order;
   public product?: Product;
@@ -22,11 +22,11 @@ OrderItem.init(
     id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
     orderId: { type: DataTypes.INTEGER, allowNull: false },
     productId: { type: DataTypes.INTEGER, allowNull: false },
-    quantity: { type: DataTypes.INTEGER, allowNull: false },
-    originalPrice: { type: DataTypes.FLOAT, allowNull: false },
-    adjustedPrice: { type: DataTypes.FLOAT, allowNull: true },
+    quantity: { type: DataTypes.FLOAT, allowNull: false }, // Total pieces (calculated)
+    originalPrice: { type: DataTypes.FLOAT, allowNull: false }, // Price per packet
+    adjustedPrice: { type: DataTypes.FLOAT, allowNull: true }, // Adjusted price per packet
     packages: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT, // Allow fractional packages
       allowNull: false,
       defaultValue: 0
     },
